# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ValueStatus {
  provider: string;
  available: boolean;
  /** @nullable */
  lastUpdated: string | null;
  entryCount: number;
}

export type ScanInputMediaType = typeof ScanInputMediaType[keyof typeof ScanInputMediaType];


export const ScanInputMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface ScanInput {
  mediaType: ScanInputMediaType;
  /**
     * @minItems 1
     * @maxItems 12
     * @items.minLength 10
     */
  frames: string[];
  /** @nullable */
  sourceName?: string | null;
}

export type PetDetectionVariant = typeof PetDetectionVariant[keyof typeof PetDetectionVariant];


export const PetDetectionVariant = {
  Normal: 'Normal',
  Neon: 'Neon',
  Mega_Neon: 'Mega Neon',
  Unknown: 'Unknown',
} as const;

export type PetDetectionPotion = typeof PetDetectionPotion[keyof typeof PetDetectionPotion];


export const PetDetectionPotion = {
  None: 'None',
  Fly: 'Fly',
  Ride: 'Ride',
  Fly_Ride: 'Fly Ride',
  Unknown: 'Unknown',
} as const;

export type PetDetectionValueStatus = typeof PetDetectionValueStatus[keyof typeof PetDetectionValueStatus];


export const PetDetectionValueStatus = {
  verified: 'verified',
  unavailable: 'unavailable',
  needs_confirmation: 'needs_confirmation',
} as const;

export interface PetDetection {
  id: string;
  petName: string;
  variant: PetDetectionVariant;
  potion: PetDetectionPotion;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  needsConfirmation: boolean;
  /** @nullable */
  frostValue: number | null;
  valueStatus: PetDetectionValueStatus;
  /** @nullable */
  evidenceFrame: number | null;
}

export interface ScanResult {
  detections: PetDetection[];
  /** @nullable */
  totalFrostValue: number | null;
  valueDataAvailable: boolean;
  valueProvider: string;
  analyzedFrames: number;
  /** @nullable */
  sourceName: string | null;
}

export type ValueLookupInputVariant = typeof ValueLookupInputVariant[keyof typeof ValueLookupInputVariant];


export const ValueLookupInputVariant = {
  Normal: 'Normal',
  Neon: 'Neon',
  Mega_Neon: 'Mega Neon',
  Unknown: 'Unknown',
} as const;

export type ValueLookupInputPotion = typeof ValueLookupInputPotion[keyof typeof ValueLookupInputPotion];


export const ValueLookupInputPotion = {
  None: 'None',
  Fly: 'Fly',
  Ride: 'Ride',
  Fly_Ride: 'Fly Ride',
  Unknown: 'Unknown',
} as const;

export interface ValueLookupInput {
  /** @minLength 1 */
  petName: string;
  variant: ValueLookupInputVariant;
  potion: ValueLookupInputPotion;
}

export type ValueLookupResultValueStatus = typeof ValueLookupResultValueStatus[keyof typeof ValueLookupResultValueStatus];


export const ValueLookupResultValueStatus = {
  verified: 'verified',
  unavailable: 'unavailable',
  needs_confirmation: 'needs_confirmation',
} as const;

export interface ValueLookupResult {
  /** @nullable */
  frostValue: number | null;
  valueStatus: ValueLookupResultValueStatus;
  provider: string;
}