# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scanner
    description: Vision scanning and verified value lookup
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /scanner/value-status:
    get:
      operationId: getValueStatus
      tags: [scanner]
      summary: Check the verified value catalog status
      responses:
        "200":
          description: Current value catalog status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValueStatus"
  /scanner/analyze:
    post:
      operationId: analyzeScannerMedia
      tags: [scanner]
      summary: Identify pets in uploaded media and calculate verified values
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "200":
          description: Scanner result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "400":
          description: Invalid input
        "502":
          description: Vision provider error
  /scanner/value-lookup:
    post:
      operationId: lookupScannerValue
      tags: [scanner]
      summary: Look up one corrected pet against the verified catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValueLookupInput"
      responses:
        "200":
          description: Verified value lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValueLookupResult"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ValueStatus:
      type: object
      properties:
        provider:
          type: string
        available:
          type: boolean
        lastUpdated:
          type: ["string", "null"]
        entryCount:
          type: integer
      required: [provider, available, lastUpdated, entryCount]
    ScanInput:
      type: object
      properties:
        mediaType:
          type: string
          enum: [image, video]
        frames:
          type: array
          minItems: 1
          maxItems: 12
          items:
            type: string
            minLength: 10
        sourceName:
          type: ["string", "null"]
      required: [mediaType, frames]
    PetDetection:
      type: object
      properties:
        id:
          type: string
        petName:
          type: string
        variant:
          type: string
          enum: [Normal, Neon, "Mega Neon", Unknown]
        potion:
          type: string
          enum: [None, Fly, Ride, "Fly Ride", Unknown]
        confidence:
          type: number
          minimum: 0
          maximum: 1
        needsConfirmation:
          type: boolean
        frostValue:
          type: ["number", "null"]
        valueStatus:
          type: string
          enum: [verified, unavailable, needs_confirmation]
        evidenceFrame:
          type: ["integer", "null"]
      required:
        [id, petName, variant, potion, confidence, needsConfirmation, frostValue, valueStatus, evidenceFrame]
    ScanResult:
      type: object
      properties:
        detections:
          type: array
          items:
            $ref: "#/components/schemas/PetDetection"
        totalFrostValue:
          type: ["number", "null"]
        valueDataAvailable:
          type: boolean
        valueProvider:
          type: string
        analyzedFrames:
          type: integer
        sourceName:
          type: ["string", "null"]
      required:
        [detections, totalFrostValue, valueDataAvailable, valueProvider, analyzedFrames, sourceName]
    ValueLookupInput:
      type: object
      properties:
        petName:
          type: string
          minLength: 1
        variant:
          type: string
          enum: [Normal, Neon, "Mega Neon", Unknown]
        potion:
          type: string
          enum: [None, Fly, Ride, "Fly Ride", Unknown]
      required: [petName, variant, potion]
    ValueLookupResult:
      type: object
      properties:
        frostValue:
          type: ["number", "null"]
        valueStatus:
          type: string
          enum: [verified, unavailable, needs_confirmation]
        provider:
          type: string
      required: [frostValue, valueStatus, provider]