# lib/api-zod/src/generated/types/petDetection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PetDetectionPotion } from './petDetectionPotion';
import type { PetDetectionValueStatus } from './petDetectionValueStatus';
import type { PetDetectionVariant } from './petDetectionVariant';

export interface PetDetection {
  id: string;
  petName: string;
  variant: PetDetectionVariant;
  potion: PetDetectionPotion;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  needsConfirmation: boolean;
  /** @nullable */
  frostValue: number | null;
  valueStatus: PetDetectionValueStatus;
  /** @nullable */
  evidenceFrame: number | null;
}