# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputMediaType } from './scanInputMediaType';

export interface ScanInput {
  mediaType: ScanInputMediaType;
  /**
     * @minItems 1
     * @maxItems 12
     * @items.minLength 10
     */
  frames: string[];
  /** @nullable */
  sourceName?: string | null;
}