# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PetDetection } from './petDetection';

export interface ScanResult {
  detections: PetDetection[];
  /** @nullable */
  totalFrostValue: number | null;
  valueDataAvailable: boolean;
  valueProvider: string;
  analyzedFrames: number;
  /** @nullable */
  sourceName: string | null;
}