# lib/api-zod/src/generated/types/valueLookupInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ValueLookupInputPotion } from './valueLookupInputPotion';
import type { ValueLookupInputVariant } from './valueLookupInputVariant';

export interface ValueLookupInput {
  /** @minLength 1 */
  petName: string;
  variant: ValueLookupInputVariant;
  potion: ValueLookupInputPotion;
}