# lib/api-zod/src/generated/types/valueLookupResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ValueLookupResultValueStatus } from './valueLookupResultValueStatus';

export interface ValueLookupResult {
  /** @nullable */
  frostValue: number | null;
  valueStatus: ValueLookupResultValueStatus;
  provider: string;
}