# lib/api-zod/src/generated/types/valueStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ValueStatus {
  provider: string;
  available: boolean;
  /** @nullable */
  lastUpdated: string | null;
  entryCount: number;
}